Add files via upload
Works with current version of Kamodo from asherp's branch, with on caveat. The line in kamodo.py with getfullargspec (in get_function_args function, currently line 271) must be replaced with the code below. Also, change line 22 from 'from inspect import getfullargspec' to 'from inspect import signature'.
```py
return symbols([a for a in signature(func).parameters if a not in hidden_args])
```